// types.ts
export interface FormState {
    title?: string;
    name: string;
    gender: string;
    age: string;
    origin: string;
    hair: string;
    clothing: string;
    skinColor: string;
    action: string;
    expression: string;
    location: string;
    time: string;
    cameraMovement: string;
    lighting: string;
    videoStyle: string;
    videoMood: string;
    sound: string;
    dialogue: string;
    additionalDetails: string;
    negativePrompt: string;
}

export interface PromptData {
    indonesian: string;
    english: string;
}